+++ src/main/java/com/campustable/be/domain/menu/dto/MenuUpdateRequest.java
@@ -3,17 +3,16 @@ package com.campustable.be.domain.menu.dto;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
 
 @Getter
 @NoArgsConstructor
 public class MenuUpdateRequest {
 
     private String menuName;
-    private BigDecimal price;
+    private Integer price;
     private String menuPicture;
     private Boolean available;
     private Integer stockQuantity;
-    //private Integer categoryId;   카테고리 바꾸는것도 필요한지 몰라서 일단 둡니다.
+//    private Long categoryId;
 
 }
